pyproject.toml
[project]
name = "qgis-plugin-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = []

[dependency-groups]
dev = [
    "pyqt5-stubs>=5.15.6.0",
    "ruff>=0.8.2",
]
